# docs/swagger.yaml
openapi: 3.0.3
info:
  title: PrescrIMed API
  description: |
    API REST completa para o sistema de gestão de prescrições médicas PrescrIMed.
    
    ## Arquitetura Multi-Tenant
    - Cada empresa possui isolamento completo de dados
    - Token JWT contém `empresaId` para filtrar automaticamente
    - Primeiro usuário cadastrado se torna administrador
    
    ## Autenticação
    Todas as rotas protegidas requerem token JWT no header:
    ```
    Authorization: Bearer {token}
    ```
    
  version: 1.0.0
  contact:
    name: Cristiano Santos
    email: cristiano.s.santos@ba.estudante.senai.br
    url: https://github.com/cristiano-superacao/prescrimed
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de Desenvolvimento
  - url: https://prescrimed-api.onrender.com/api
    description: Servidor de Produção
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido no login ou registro
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: Mensagem de erro
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              location:
                type: string
    Usuario:
      type: object
      properties:
        id:
          type: string
          example: 64abc123def456789
        nome:
          type: string
          example: Dr. João Silva
        email:
          type: string
          format: email
          example: joao.silva@clinica.com
        telefone:
          type: string
          example: (11) 98765-4321
        role:
          type: string
          enum: [admin, usuario]
          example: admin
        permissoes:
          type: array
          items:
            type: string
          example: [dashboard, pacientes, prescricoes, usuarios]
        empresaId:
          type: string
          example: 64xyz789abc012345
        empresaNome:
          type: string
          example: Clínica Saúde Total
        ativo:
          type: boolean
          example: true
        criadoEm:
          type: string
          format: date-time
        ultimoAcesso:
          type: string
          format: date-time
...
    Paciente:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
          example: Maria Santos
        cpf:
          type: string
          example: 123.456.789-00
        dataNascimento:
          type: string
          format: date
          example: 1985-05-15
        sexo:
          type: string
          enum: [M, F, Outro]
          example: F
        telefone:
          type: string
          example: (11) 91234-5678
        email:
          type: string
          example: maria.santos@email.com
        endereco:
          type: object
        tipoSanguineo:
          type: string
          example: O+
        peso:
          type: number
          example: 65.5
        altura:
          type: number
          example: 1.65
        alergias:
          type: array
          items:
            type: string
          example: [Dipirona, Penicilina]
        condicoesMedicas:
          type: array
          items:
            type: string
          example: [Hipertensão, Diabetes]
        medicamentosUso:
          type: array
          items:
            type: string
          example: [Losartana 50mg, Metformina 850mg]
        convenio:
          type: object
          properties:
            nome:
              type: string
            numeroCarteira:
              type: string
            validade:
              type: string
              format: date
        contatoEmergencia:
          type: object
          properties:
            nome:
              type: string
            telefone:
              type: string
            parentesco:
              type: string
        empresaId:
          type: string
        ativo:
          type: boolean
    Prescricao:
      type: object
      properties:
        id:
          type: string
        pacienteId:
          type: string
        paciente:
          type: object
          properties:
            nome:
              type: string
            cpf:
              type: string
        medicoId:
          type: string
        medico:
          type: object
          properties:
            nome:
              type: string
            crm:
              type: string
        medicamentos:
          type: array
          items:
            type: object
            properties:
              nome:
                type: string
                example: Amoxicilina
              dosagem:
                type: string
                example: 500mg
              quantidade:
                type: number
                example: 21
              posologia:
                type: string
                example: Tomar 1 comprimido a cada 8 horas
              viaAdministracao:
                type: string
                example: Oral
              duracao:
                type: string
                example: 7 dias
        tipo:
          type: string
          enum: [comum, controlado, amarelo, azul]
          example: comum
        status:
          type: string
          enum: [ativa, cancelada, arquivada]
          example: ativa
        observacoes:
          type: string
        dataEmissao:
          type: string
          format: date-time
        empresaId:
          type: string
paths:
  /auth/register:
    post:
      tags:
        - Autenticação
      summary: Registrar nova empresa e administrador
      description: Cria uma nova empresa e o primeiro usuário (administrador) simultaneamente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nomeEmpresa
                - email
                - senha
                - nomeAdmin
              properties:
                nomeEmpresa:
                  type: string
                  example: Clínica Saúde Total
                cnpj:
                  type: string
                  example: 12.345.678/0001-90
                email:
                  type: string
                  format: email
                  example: admin@clinica.com
                senha:
                  type: string
                  format: password
                  minLength: 6
                  example: senha123
                nomeAdmin:
                  type: string
                  example: Dr. João Silva
                telefone:
                  type: string
                  example: (11) 98765-4321
      responses:
        '201':
          description: Empresa e usuário criados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  usuario:
                    $ref: '#/components/schemas/Usuario'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email ou CNPJ já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Autenticação
      summary: Login de usuário
      description: Autentica um usuário e retorna token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - senha
              properties:
                email:
                  type: string
                  format: email
                  example: admin@clinica.com
                senha:
                  type: string
                  format: password
                  example: senha123
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  usuario:
                    $ref: '#/components/schemas/Usuario'
        '400':
          description: Dados inválidos
        '401':
          description: Credenciais inválidas
  /auth/refresh:
    post:
      tags:
        - Autenticação
      summary: Renovar token JWT
      description: Gera um novo token usando o refresh token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token renovado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Token inválido ou expirado
...
  /pacientes:
    get:
      tags:
        - Pacientes
      summary: Listar pacientes
      description: Lista pacientes da empresa com busca opcional
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Buscar por nome ou CPF
      responses:
        '200':
          description: Lista de pacientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Paciente'
    
    post:
      tags:
        - Pacientes
      summary: Cadastrar paciente
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - cpf
                - dataNascimento
              properties:
                nome:
                  type: string
                cpf:
                  type: string
                dataNascimento:
                  type: string
                  format: date
                sexo:
                  type: string
                telefone:
                  type: string
                email:
                  type: string
                endereco:
                  type: object
                tipoSanguineo:
                  type: string
                peso:
                  type: number
                altura:
                  type: number
                alergias:
                  type: array
                  items:
                    type: string
                condicoesMedicas:
                  type: array
                  items:
                    type: string
                medicamentosUso:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Paciente cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paciente'
        '400':
          description: Dados inválidos
        '409':
          description: CPF já cadastrado
  /pacientes/{id}:
    get:
      tags:
        - Pacientes
      summary: Buscar paciente por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados do paciente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paciente'
        '404':
          description: Paciente não encontrado
    
    put:
      tags:
        - Pacientes
      summary: Atualizar paciente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Paciente'
      responses:
        '200':
          description: Paciente atualizado
        '404':
          description: Paciente não encontrado
    
    delete:
      tags:
        - Pacientes
      summary: Desativar paciente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Paciente desativado
        '404':
          description: Paciente não encontrado
...
  /prescricoes/paciente/{id}:
    get:
      tags:
        - Prescrições
      summary: Listar prescrições de um paciente
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID do paciente
      responses:
        '200':
          description: Prescrições do paciente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prescricao'
...